型の整合性を改善し、メソッドを更新
`Texture.h` で `setImage` メソッドの呼び出しにおいて、`getWidth(level)` と `getHeight(level)` の戻り値を `static_cast<uint32_t>` でキャストしました。これにより、型の不一致を防ぎます。

`Material.cpp` では、`setParameter` メソッドのシグネチャを変更し、テクスチャのサンプラーを設定する機能を強化しました。

`Texture.cpp` では、`getWidth`、`getHeight`、および `getDepth` メソッドの `level` 引数を `static_cast<uint8_t>` でキャストするように変更し、関数の動作を安定させました。

+++ engine/include/ocf/renderer/Texture.h
@@ -49,7 +49,8 @@ public:
 
     void setImage(size_t level, PixelBufferDescriptor&& buffer) const
     {
-        setImage(level, 0, 0, 0, getWidth(level), getHeight(level), m_depth, std::move(buffer));
+        setImage(level, 0, 0, 0, static_cast<uint32_t>(getWidth(level)),
+                 static_cast<uint32_t>(getHeight(level)), m_depth, std::move(buffer));
     }
 
 private:
